+++ extensions/umami/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Umami Changelog
 
+## [Fix Website Stats Always Zero in Cloud] - 2026-01-25
+
+- Fix: Website Stats are now correctly displayed when using Cloud V3
+
 ## [View Websites (Menu Bar)] - 2026-01-24
 
 - View Websites (Menu Bar)